# swagger.yaml
# @format

openapi: 3.0.0
info:
  title: Item API
  description: API for managing items
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /items:
    get:
      summary: Get all items
      responses:
        200:
          description: A list of items
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
    post:
      summary: Create a new item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        201:
          description: Item created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string

  /items/{id}:
    put:
      summary: Update an existing item
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the item to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        200:
          description: The updated item
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
        404:
          description: Item not found

    delete:
      summary: Delete an item
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the item to delete
          schema:
            type: integer
      responses:
        204:
          description: Item deleted